Split resolver from DNS (protocol) module
