+++ .hooks/commit-msg
@@ -4,7 +4,7 @@ grep -ve '^#' $1 >$TMP
 cat $TMP >/tmp/dumpmsg
 
 abort() {
-    echo "Re-edit with:\n\ngit ci --amend -F .git/COMMIT_EDITMSG -e"
+    echo "Re-edit with:\n\n<last_git_command> -F .git/COMMIT_EDITMSG -e"
     exit 1
 }
 
@@ -13,8 +13,8 @@ if [ "$(head -n 1 $TMP | wc -c)" -gt 50 ]; then
     abort
 fi
 
-if [ "$(head -n 1 $TMP | tail -c -2)" != "." ]; then
-    echo "Commit summary should end with period"
+if [ "$(head -n 1 $TMP | tail -c -2)" == "." ]; then
+    echo "Commit summary should not end with period"
     abort
 fi
