+++ modules/dns.h
@@ -17,32 +17,6 @@ extern "C" {
 #define NS_DNS_AAAA_RECORD  0x1c  /* Lookup IPv6 address */
 #define NS_DNS_MX_RECORD    0x0f  /* Lookup mail server for domain */
 
-struct ns_dns_message;
-
-typedef void (*ns_resolve_callback_t)(struct ns_dns_message *, void *);
-
-/* Options for `ns_resolve_async_opt`. */
-struct ns_resolve_async_opts {
-  const char *nameserver_url;
-  int max_retries;    /* defaults to 2 if zero */
-  int timeout;        /* in seconds; defaults to 5 if zero */
-  int accept_literal; /* pseudo-resolve literal ipv4 and ipv6 addrs */
-  int only_literal;   /* only resolves literal addrs; sync cb invocation */
-};
-
-int ns_resolve_async(struct ns_mgr *mgr, const char *, int,
-                   ns_resolve_callback_t, void *);
-int ns_resolve_async_opt(struct ns_mgr *mgr, const char *, int,
-                       ns_resolve_callback_t, void *,
-                       struct ns_resolve_async_opts opts);
-
-struct ns_dns_resource_record *ns_dns_next_record(
-    struct ns_dns_message *, int, struct ns_dns_resource_record *);
-
-int ns_dns_parse_record_data(struct ns_dns_message *,
-                             struct ns_dns_resource_record *, void *, size_t);
-
-/* low-level */
 #define NS_MAX_DNS_QUESTIONS 32
 #define NS_MAX_DNS_ANSWERS   32
 
@@ -63,6 +37,12 @@ struct ns_dns_message {
   struct ns_dns_resource_record questions[NS_MAX_DNS_QUESTIONS];
   struct ns_dns_resource_record answers[NS_MAX_DNS_ANSWERS];
 };
+
+struct ns_dns_resource_record *ns_dns_next_record(
+    struct ns_dns_message *, int, struct ns_dns_resource_record *);
+
+int ns_dns_parse_record_data(struct ns_dns_message *,
+                             struct ns_dns_resource_record *, void *, size_t);
 
 void ns_send_dns_query(struct ns_connection*, const char *, int);
 int ns_parse_dns(const char *, int, struct ns_dns_message *);
