+++ modules/resolv-internal.h
@@ -5,4 +5,4 @@
 
 /* internals that need to be accessible in unit tests */
 
-NS_INTERNAL int ns_dns_resolve_hosts(const char *, struct in_addr *);
+NS_INTERNAL int ns_resolve_etc_hosts(const char *, struct in_addr *);
